h3e
Made more changes in migrating from using own collision detection and movement to using the Rigidbody based versions. Need to fix movement so that it is directed in the cross product between the right-side of the player transform and the ground: this will cause it to go in the direction they are facing at an angle, causing them to not continually fall of and not on a downwards slope.