Now equipping works again :D + other
General:
Equipping items now work again. Equipped items will also display an E-icon on their icon in the inventory.
TODO: Setup so that when equipping a new item, EquipmentManager will call update on the inventory, making it display the E-icon on the newly equipped item. Also make it so that the E-icon doesn't get displayed when you switch to another tab. ALSO: Fix so that the items get equipped in the correct slot. Fix that by looking into EquipmentManager, and making a way to tie an equipment into a slot, except if its a weapon, then you have to specify which hand and which slot of the hand (1-3).