More progress
Made so that effects can be applied from the outside at least. Currently you have to create the effects and drag them in via the inspector, I need to change it so you can just choose an enum regarding which effect you want, and then it will be applied from a set of premade ones, unless you want a custom one, in which case Idk what to do yet.

// TODO:: Change this so you can choose between specifying the power per tick or the total power over the duration