Started adding the roll functionality
Didn't have much time as usual, tested out some numbers in order to make the TotalDistance of the slide and roll actually be representative of how far you travel. A problem is that both the travel distance and the duration of the travel has to be taken into account when adding force to the player; as the force should be faster in the beginning and slower at the end. I'll probably finish it tomorrow, or at least get the actual roll animation in so I can tell if it needs root motion or not.